backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
Pillow==10.1.0
python-dotenv==1.0.0
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
markdown==3.5.1
django-taggit==5.0.1
